handle incorrect ids

+++ routes/api.js
@@ -5,6 +5,10 @@ var users = require('./api/user');
 var projects = require('./api/project');
 var auth = require('../controllers/auth');
 
+function isIdValid(id) {
+  return id.match(/^[0-9a-fA-F]{24}$/);
+}
+
 router.route('/authenticate')
   .get(auth.authenticate, function(req, res, next) {
     res.sendStatus(200);
@@ -22,12 +26,15 @@ router.route('/users')
 /* Single user routes */
 router.route('/users/:user_id')
   .get(function(req, res, next) {
+    if (!isIdValid(req.params.user_id)) return res.sendStatus(404);
     users.getUser(req, res, req.params.user_id, next)
   })
   .put(auth.authenticate, function(req, res, next) {
+    if (!isIdValid(req.params.user_id)) return res.sendStatus(404);
     users.updateUser(req, res, req.params.user_id, next)
   })
   .delete(auth.authenticate, function(req, res, next) {
+    if (!isIdValid(req.params.user_id)) return res.sendStatus(404);
     users.deleteUser(req, res, req.params.user_id, next)
   });
 
@@ -43,22 +50,27 @@ router.route('/projects')
 /* Single project routes */
 router.route('/projects/:project_id')
   .get(auth.authenticate, function(req, res, next) {
+    if (!isIdValid(req.params.project_id)) return res.sendStatus(404);
     projects.getProject(req, res, req.params.project_id, next)
   })
   .put(auth.authenticate, function(req, res, next) {
+    if (!isIdValid(req.params.project_id)) return res.sendStatus(404);
     projects.updateProject(req, res, req.params.project_id, next)
   })
   .delete(auth.authenticate, function(req, res, next) {
+    if (!isIdValid(req.params.project_id)) return res.sendStatus(404);
     projects.deleteProject(req, res, req.params.project_id, next)
   });
 
 router.route('/projects/:project_id/join')
   .post(auth.authenticate, function(req, res, next) {
+    if (!isIdValid(req.params.project_id)) return res.sendStatus(404);
     projects.joinProject(req, res, req.params.project_id, next)
   });
 
 router.route('/projects/:project_id/leave')
   .delete(auth.authenticate, function(req,res, next) {
+    if (!isIdValid(req.params.project_id)) return res.sendStatus(404);
     projects.leaveProject(req, res, req.params.project_id, next)
   });
